add getter_setter

+++ philo/src/time.c
@@ -1,8 +1,20 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   time.c                                             :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: kfukuhar <kfukuhar@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/09/26 15:11:16 by kfukuhar          #+#    #+#             */
+/*   Updated: 2024/09/26 15:11:16 by kfukuhar         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "philo.h"
 
-long long int	current_timestamp(void)
+long int	current_timestamp(void)
 {
-	struct timeval	te;
+	struct timeval te;
 
 	gettimeofday(&te, NULL);
 	return (te.tv_sec * 1000LL + te.tv_usec / 1000);
